Refactor Arguments class
