--- bot/configuration/cli_options.py
@@ -1,53 +1,0 @@
-"""
-Definition of cli arguments used in arguments.py
-"""
-import argparse
-
-from bot import __version__, constants
-
-
-def check_int_positive(value: str) -> int:
-    try:
-        uint = int(value)
-        if uint <= 0:
-            raise ValueError
-    except ValueError:
-        raise argparse.ArgumentTypeError(
-            f"{value} is invalid for this parameter, should be a positive integer value"
-        )
-    return uint
-
-
-def check_int_nonzero(value: str) -> int:
-    try:
-        uint = int(value)
-        if uint == 0:
-            raise ValueError
-    except ValueError:
-        raise argparse.ArgumentTypeError(
-            f"{value} is invalid for this parameter, should be a non-zero integer value"
-        )
-    return uint
-
-
-class Arg:
-    # Optional CLI arguments
-    def __init__(self, *args, **kwargs):
-        self.cli = args
-        self.kwargs = kwargs
-
-
-# List of available command line options
-AVAILABLE_CLI_OPTIONS = {
-    "version": Arg(
-        '-V', '--version',
-        action='version',
-        version=f'%(prog)s {__version__}',
-    ),
-    "config": Arg(
-        '-c', '--config',
-        help=f'Specify configuration file (default: `{constants.DEFAULT_CONFIG}`). ',
-        action='append',
-        metavar='PATH',
-    ),
-}
